separate BlockSimulation as a module project
