+++ NineChronicles.Modules.BlockSimulation/MockWorldState.cs
@@ -9,7 +9,7 @@ using Libplanet.Store.Trie;
 using Libplanet.Types.Assets;
 using Libplanet.Types.Consensus;
 
-namespace NineChronicles.Mods.Athena.BlockSimulation
+namespace NineChronicles.Modules.BlockSimulation
 {
 
     public class MockWorldState : IWorldState
